@@ -1,24 +1,24 @@
 module github.com/mashiike/gdnotify
 
-go 1.18
+go 1.20
 
 require (
 	github.com/Songmu/flextime v0.1.0
-	github.com/aws/aws-lambda-go v1.32.0
-	github.com/aws/aws-sdk-go-v2 v1.16.5
-	github.com/aws/aws-sdk-go-v2/config v1.15.10
-	github.com/aws/aws-sdk-go-v2/feature/s3/manager v1.11.15
-	github.com/aws/aws-sdk-go-v2/service/dynamodb v1.15.6
-	github.com/aws/aws-sdk-go-v2/service/eventbridge v1.16.3
-	github.com/aws/aws-sdk-go-v2/service/s3 v1.26.11
-	github.com/aws/aws-sdk-go-v2/service/ssm v1.27.2
-	github.com/aws/smithy-go v1.11.3
+	github.com/aws/aws-lambda-go v1.37.0
+	github.com/aws/aws-sdk-go-v2 v1.17.5
+	github.com/aws/aws-sdk-go-v2/config v1.18.15
+	github.com/aws/aws-sdk-go-v2/feature/s3/manager v1.11.55
+	github.com/aws/aws-sdk-go-v2/service/dynamodb v1.18.5
+	github.com/aws/aws-sdk-go-v2/service/eventbridge v1.18.5
+	github.com/aws/aws-sdk-go-v2/service/s3 v1.30.5
+	github.com/aws/aws-sdk-go-v2/service/ssm v1.35.5
+	github.com/aws/smithy-go v1.13.5
 	github.com/fatih/color v1.13.0
 	github.com/fujiwara/logutils v1.1.0
 	github.com/fujiwara/ridge v0.6.1
 	github.com/gofrs/flock v0.8.1
 	github.com/google/uuid v1.3.0
-	github.com/hashicorp/go-version v1.5.0
+	github.com/hashicorp/go-version v1.6.0
 	github.com/kayac/go-config v0.6.0
 	github.com/ken39arg/go-flagx v0.0.0-20220608183922-7cf7c6c0093c
 	github.com/mashiike/didumean v0.1.2
